requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==2.0.12
cloudinary==1.36.0
crispy-bootstrap5==0.7
dj-database-url==2.1.0
Django==4.2.8
django-cloudinary-storage==0.3.0
django-crispy-forms==2.0
gunicorn==21.2.0
idna==3.10
Pillow==10.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2025.2
requests==2.31.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==1.26.20
whitenoise==6.6.0
